feat: dataset merge script

+++ tools/dataset/merge_dataset.sh
@@ -2,9 +2,10 @@
 
 set -e
 
-INCLUDE_REDACTED=true
+INCLUDE_REDACTED=false
 FILTERD_SCORE=7
-NEXT_TOKEN_PERCENT=0.25
+NEXT_TOKEN_PERCENT=0.5
+USE_OPEN_ASSISTANT=false
 
 OUTPUT_DIR=/bb/llm/gaf51275/datasets/raw/instruct/training/exp2-filtered-$FILTERD_SCORE-next_token-$NEXT_TOKEN_PERCENT
 
@@ -12,32 +13,42 @@ if $INCLUDE_REDACTED; then
   OUTPUT_DIR=$OUTPUT_DIR-redacted
 fi
 
+if ! $USE_OPEN_ASSISTANT; then
+  OUTPUT_DIR=$OUTPUT_DIR-no-oasst
+fi
+
 mkdir -p $OUTPUT_DIR
 
-FILES=(
-  "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-33k-ja/lm_filtered_split_1.jsonl"
-  "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-33k-ja/lm_filtered_split_2.jsonl"
-  "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-33k-ja/lm_filtered_split_3.jsonl"
-  "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-33k-ja/lm_filtered_split_4.jsonl"
-  "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-top1-en-chat-sft/data/train.jsonl"
-)
+if $USE_OPEN_ASSISTANT; then
+  FILES=(
+    "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-33k-ja/lm_filtered_split_1.jsonl"
+    "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-33k-ja/lm_filtered_split_2.jsonl"
+    "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-33k-ja/lm_filtered_split_3.jsonl"
+    "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-33k-ja/lm_filtered_split_4.jsonl"
+    "/bb/llm/gaf51275/datasets/raw/instruct/general/oasst2-top1-en-chat-sft/data/train.jsonl"
+  )
 
-MERGED_FILE=$OUTPUT_DIR/merged.jsonl
+  MERGED_FILE=$OUTPUT_DIR/merged.jsonl
 
-for FILE in "${FILES[@]}"; do
-  cat $FILE >> $MERGED_FILE
-done
+  for FILE in "${FILES[@]}"; do
+    cat $FILE >> $MERGED_FILE
+  done
 
-# fileter
-python tools/dataset/fileter.py \
-  --input_file $MERGED_FILE \
-  --output_file $OUTPUT_DIR/train.jsonl \
-  --threshold $FILTERD_SCORE
+  # filter
+  python tools/dataset/fileter.py \
+    --input_file $MERGED_FILE \
+    --output_file $OUTPUT_DIR/train.jsonl \
+    --threshold $FILTERD_SCORE
 
-rm $MERGED_FILE
+  rm $MERGED_FILE
 
-echo "Filtered open assistant data:"
-wc -l $OUTPUT_DIR/train.jsonl
+  echo "Filtered open assistant data:"
+  wc -l $OUTPUT_DIR/train.jsonl
+else
+  # Open Assistant データを使用しない場合は空のファイルを作成
+  touch $OUTPUT_DIR/train.jsonl
+  echo "Skipped Open Assistant data processing."
+fi
 
 if $INCLUDE_REDACTED; then
   LMSYS_FILE=/bb/llm/gaf51275/datasets/raw/instruct/lmsys-chat-1m/sft/lmsys-chat-1m-train.jsonl
